server/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
mysqlclient
python-dotenv==1.0.1
bcrypt
PyJWT==2.10.1